fix: 게시글 C 오류 수정
