+++ src/types/props.ts
@@ -23,6 +23,7 @@ export interface WriteProps {
   writeModal: boolean;
   setWriteModal: React.Dispatch<SetStateAction<boolean>>;
   setSearchModal: React.Dispatch<SetStateAction<boolean>>;
+  setPost: React.Dispatch<SetStateAction<Post | null>>;
 }
 
 export interface PostsProps {
